Remove priority queue from RhtPQMap and entire project
