+++ yorkie/src/main/kotlin/dev/yorkie/document/Document.kt
@@ -10,7 +10,7 @@ import dev.yorkie.document.change.ChangePack
 import dev.yorkie.document.change.CheckPoint
 import dev.yorkie.document.crdt.CrdtObject
 import dev.yorkie.document.crdt.CrdtRoot
-import dev.yorkie.document.crdt.RhtPQMap
+import dev.yorkie.document.crdt.ElementRht
 import dev.yorkie.document.json.JsonObject
 import dev.yorkie.document.time.ActorID
 import dev.yorkie.document.time.TimeTicket
@@ -40,7 +40,7 @@ public class Document private constructor(
     private val localChanges = mutableListOf<Change>()
 
     @Volatile
-    private var root: CrdtRoot = CrdtRoot(CrdtObject(InitialTimeTicket, rht = RhtPQMap()))
+    private var root: CrdtRoot = CrdtRoot(CrdtObject(InitialTimeTicket, rht = ElementRht()))
 
     @get:VisibleForTesting
     @Volatile
